Add scripts and output for determining specificity of IAV guides
Cameron ran diagnostic-guide-design separately on each IAV subtype. As a quick
way to use the outputs to perform differential identification of IAV subtypes,
this takes the guide sets output by each of the runs (2k8_H_uguides.fasta),
uses bwa to align them back to all of the IAV sequences, and counts the
fraction of sequences for each subtype that each guide and guide set hits.
The output of calculate_subtype_coverage.sh is in out/.