Design guides for IAV subtype identification using the differential identification feature added to diagnostic-guide-design
This includes a script, run.sh, that runs commands to generate guides
separately for H and N subtypes. For each, it designs them completely
independently ('no-id') and also using the differential identification (--id)
option added to the diagnostic-guide-design software ('id').

This uses commit e5346d9 on branch diff-id of the diagnostic-guide-design repo.