Add table giving year for each N subtype sequence
This adds a script to map sequence name (in a particular format) to the year of
the sequence, printing out a tsv. This commit has the output of that script,
run on all N subtype sequences (Hstar), saved to
2k8_Hstar_aligned_amplicon.seq-with-year.tsv.  The script output 12 'unknown's
(sequences for which it could not determine the year). I manually replaced
'unknown' with the year in the file for those 12 sequences, and also manually
fixed incorrect years that were output (well outside the 2008-2018 range) for 2
sequences.