Update table giving year for each N subtype sequence and add table for each H subtype
This also adds a table giving the year for each H/N accession.

I manually adjusted the years in the tables to replace 'unknown' with
an actual year (by manually inspecting the sequence names) and to fix
obviously incorrect years that were output (well outside the 2008-2018
range).

This is based on updated sequence data that Cameron used as input to
guide design. This updates map_seq_name_to_year.sh because now the
sequence headers have spaces rather than underscores.