Calculate per-year coverage for IAV subtyping on a per-guide-set basis
Previously, the table showing coverage per-year for each IAV subtype was
calculated by combining all guides across all guide sets. This updates the
script that produces the table to still produce such a table, but also to
produce pairwise tables: one for each combination of subtype and guide set,
which shows the coverage of that subtype per-year provided by that guide set.

This commit also contains the output of that script for the IAV subtyping v3
design.