Re-calculate coverage data for N subtype guides using new target IAV sequences
The previous coverage data was calculated by aligning against a version of the
IAV N subtype sequences (Hstar) that was *not* what was used in the guide
design for the N subtype guides. Thus, some guide sets did not achieve the
coverage that they should have been guaranteed to meet in the design. This
commit updates the IAV N subtype sequences to match what was used in the guide
design, and also includes the output of the coverage calculation scripts using
these updated sequences.