Create initial list of species for pan-viral ID design
This is based on a download of NCBI's viral accession list
on 2018-09-01 (taxid10239.nbr.20180901.txt). The file
human-host-lineages-to-add.txt lists lineages to explicitly
include in the list, based on my experience with hybsel
design (it includes lineages that, as some point, were
not labeled as having human as a host but probably should
have been). The output of running make_species_list.sh is
species-list.tsv.