Make small modifications to 2018-09-05 viral accession list
This makes the same changes as the ones made to the 2018-09-01
viral accession list in commit 670fba (plus one additional change).

In particular:

This renames the sequence with species 'unidentified circular ssDNA virus'
to have species 'circular ssDNA virus HV-CV1'.

This also corrects the lineage of two sequences that were likely misclassified:
  - Change a Bovine calicivirus sequence to be classified as:
    'Caliciviridae,Vesivirus,Vesicular exanthema of swine virus'
  - Change a Mamastrovirus 1 sequence to be classified as:
    'Astroviridae,Mamastrovirus,Mamastrovirus 1'

In addition, it makes the following change not made in 670fba:
  - Change two Morogoro mammarenavirus sequences to be classified as:
    'Arenaviridae,Mammarenavirus,Morogoro mammarenavirus'