Add sequences that were not in the viral accession list but should be
These are listed in the spreadsheet at:
  https://docs.google.com/spreadsheets/d/1DFP2Qc8VmcUs-OIzBupHYAovSNrF8yk5T3IY6AnQELY/edit#gid=214390472

These sequences are from species with cases of human infection, but there
are currently no sequences from these species in the viral accession
list pulled on 2018-09-05. In most cases, this is probably because
the species does not (yet) have a RefSeq and thus there are no genome
neighbors for them.