Be more robust to incomplete NCBI results when downloading
Sometimes the results of a query do not included all the requested
accessions in that query, which leads to errors later on. In the
past I had just re-run download_seqs.py until it did not encounter
such an error (usually the second try worked). This checks results
immediately after each query, and re-tries the query if its results
do not include all queried accessions.