Update download_seqs.py to include a custom list of Influenza sequences
Because there are so many Influenza A/B/C sequences, these are not
comprehensively covered in the viral genome accession list and are
instead in a separate NCBI database. To include these sequences,
this commit adds an option to download_seqs.py to accept them in
a separate table. It uses these sequences both in generating the
species list and making fasta files.