Update IAV subtype coverage tables on v1 and v2 designs
This commit has re-run coverage calculations for the v1
and v2 designs using the original alignment score (as determined
by bwa) (*.orig-score.*) as well as a recomputed score (*.recomputed-score.*)
that accounts for G-U pairing.

See 081a81f for more details on the recomputed score and the
new tables on the v3 designs.

This includes re-runs on both the H and N subtypes for the v2 design,
but the v1 design was only on the H subtype.