Make changes to pan-viral design input based on alignment issues
This makes several changes to the input accessions used for
pan-viral guide design (2018-09) based on issues in the sequences
that Cameron noticed when making alignments from the fasta files.

The changes to the accession list (taxid10239.nbr.20180905.modified.txt) are:
  - Remove all Giant panda anellovirus sequences; these do not
    appear to infect humans
  - Remove 1010 other sequences for various reasons (e.g., permuted
    genome)
  - Change the segment of KU978811.1 (an LCMV sequence) from S to L

This also adds a list of sequences -- seqs-to-reverse-complement.txt --
for which the reverse complement should be used (i.e., the sequence
was in the wrong orientation for alignment).

See
  https://docs.google.com/spreadsheets/d/1D0wLiIVuvjz0Dvp67r2CkSe-La0cqBxN4MtjQC_vTEQ/edit
for more details.