Make changes to pan-viral design input
This removes two sequences (NC_038721 and NC_038722) corresponding
to species for which we do not have the complete genome (i.e.,
only one segment of a segmented virus); this effectively removes the
species.

This also assigns a segment label to 83 accessions that should
have had a labeled segment but do not.

See
  https://docs.google.com/spreadsheets/d/1D0wLiIVuvjz0Dvp67r2CkSe-La0cqBxN4MtjQC_vTEQ/edit
for more details.