Fix issue in compare-to-baseline analysis in which coordinates differ across resamplings
Previously, each of the resamplings was performed by `design.py` using
`--sample-seqs`. This produces (and outputs) a separate alignment from
each resampling. This was a mistake because the different alignments can
be in (probably slightly) different coordinate spaces.  The outputs of
`design.py` and `design_naively.py` include window positions, and so
when we match these up across the different runs they may not be
comparable.

This commit fixes the issue by producing an alignment just once (in
`make_designs.sh`) and then resampling the alignment to create a fasta
file as input for each design. By sampling the same alignment, all
designs are produced across the same coordinate space and so window
positions across designs can be matched.