Compute coverage against a group of years for which there are few sequences
For the change-over-time analyses, this would normally compute
coverage against each year in a range [A, B] for each design. But
some of the years (especially more distant ones -- i.e., near A) may
have few sequences available and so it is hard to draw conclusions about
coverage against those years or plot them. Therefore, this determines a
group of years [A, A'] (where the year A'+1 has a sufficient number of
sequences) and computes coverage against all the sequences in [A, A'].

This updates the scripts to compute these coverages, and also the
results of running them.