Benchmark approach taking signature of full 28-mer (not splitting)
This modifies benchmark_trie_signatures.py to implement an approach
that shards 28-mers into tries based on a signature of the full
28-mer; this is the 'full' approach. It still benchmarks this against
the previous approach ('split'), which splits each 28-mer into two
14-mers and computes a signature for each 14-mer.

This commit includes the output of running benchmark_trie_signatures.py
(on a fraction 0.0001, 0.0016, and 0.0128 of the full dataset of
28-mers).

This commit also removes prior output files on other subsamplings
which were not run using the updated benchmark_trie_signatures.py. The
removed output files only contain results on the 'split' approach and
not on the 'full' approach.