Add analysis of large detection design panel
This includes the summary design results from a panel across all viral
species with >= 10 sequences (without concern for specificity); it was
run in my `continuous-dgd` repo. The summary information, copied from
that repo, is in data/summary.one-segment-per-taxid.tsv.

This commit also includes a simplification of that data, a script to
produce plots from it, and the produced plots (.pdf files).