Add redesign of flavivirus panel with new parameters
This moves the old design from out/ to out_PM2_N5/. The main change is
that the new design uses (PM (number of tolerated primer mismatches) =
3), whereas the old design uses PM=2. Note that the large
all-viral-with-ge10-seqs design uses PM=3, so using PM=3 enables
comparison with that design. However, in the all-viral-with-ge10-seqs
design, we used PM=2 for dengue virus; thus, it is important to keep the
old (PM=2) design as well.

Lowering N (the number of designs to find) from 5 to 2 in the new design
should not matter, because we only use the best (lowest cost) design in
plots/analysis. However, it does improve runtime.