Remove influenza A virus from design analyses
Previously, the influenza neighbors fetched were only from the viral
neighbors list, not from NCBI's influenza virus database, and thus only
used a very small fraction of all influenza neighbors. This is fixed in
fe14626.

However, due to the sheer number of influenza A virus neighbors,
including them in these analyses is difficult because these analyses
have many replicates -- it is slow to run. If they are to be included,
it might make sense to only include recent neighbors (e.g., last 5
years).