Update all-vertebrate design results data
This includes data for 1,933 vertebrate-infecting species. 7 species did
not yield designs with the "specific_max-activity" objective (so, with
that objective, this only includes 1,926 species); all of these species
have only 1 genome sequence. There are many other species that could not
produce a design with the "*_min-guides" objective, likely because no
guides could be found that had sufficiently high activity and achieve the
desired coverage (and, in the "specific_*" objective, are specific).