Fix bug when reading all-vertebrate design results
By default, read.table() sets `quote="\"'"`, which means that everything
between quotes is read as a single field of data. Some taxa names have
quotes in them, and thus read.table() was not correctly reading the data
(and only a fraction of all the rows in design-results.tsv). This commit
fixes the problem by setting `quote=""` to disable quoting.